Convert twitch api to gql, fix clip downloading functionality
