+++ internal/config/config.go
@@ -1,7 +1,6 @@
 package config
 
 import (
-	"github.com/skhanal5/clip-farmer/internal/model/twitch"
 	"log"
 	"os"
 )
@@ -10,8 +9,7 @@ type Config struct {
 	TiktokClientKey    string
 	TikTokClientSecret string
 	TwitchClientId     string
-	TwitchClientSecret string
-	TwitchOAuthConfig  twitch.TwitchOAuthResponse
+	TwitchClientOAuth  string
 }
 
 func NewConfig() Config {
@@ -20,11 +18,6 @@ func NewConfig() Config {
 		os.Getenv("TIKTOK_CLIENT_KEY"),
 		os.Getenv("TIKTOK_CLIENT_SECRET"),
 		os.Getenv("TWITCH_CLIENT_ID"),
-		os.Getenv("TWITCH_CLIENT_SECRET"),
-		twitch.TwitchOAuthResponse{},
+		os.Getenv("TWITCH_CLIENT_OAUTH"),
 	}
-}
-
-func (config *Config) SetTwitchBearerToken(oauthResponse twitch.TwitchOAuthResponse) {
-	config.TwitchOAuthConfig = oauthResponse
 }
