+++ internal/model/tiktok/tiktok_oauth_request_body.go
@@ -1,4 +1,4 @@
-package tiktok
+package model
 
 type TikTokOAuthRequestBody struct {
 	ClientKey    string `json:"client_key"`
