+++ internal/model/twitch/twitch_gql_user.go
@@ -1,4 +1,4 @@
-package twitch
+package model
 
 type UserResponse struct {
 	Data       UserData   `json:"data"`
