--- internal/model/twitch/twitch_oauth_response.go
@@ -1,7 +1,0 @@
-package twitch
-
-type TwitchOAuthResponse struct {
-	AccessToken string `json:"access_token"`
-	ExpiresIn   int    `json:"expires_in"`
-	TokenType   string `json:"token_type"`
-}
